+++ packages/amplify-ui-components/src/components/amplify-confirm-sign-up/amplify-confirm-sign-up.tsx
@@ -129,7 +129,10 @@ export class AmplifyConfirmSignUp {
         secondaryFooterContent={
           <div>
             <span>
-              <amplify-button variant="anchor" onClick={() => this.handleAuthStateChange(AuthState.SignIn)}>
+              <amplify-button
+                variant="anchor"
+                onClick={() => this.handleAuthStateChange(AuthState.SignIn)}
+              >
                 {I18n.get(Translations.BACK_TO_SIGN_IN)}
               </amplify-button>
             </span>