+++ packages/amplify-ui-components/src/components/amplify-federated-buttons/amplify-federated-buttons.tsx
@@ -13,7 +13,7 @@ export class AmplifyFederatedButtons {
   /** The current authentication state. */
   @Prop() authState: AuthState = AuthState.SignIn;
   /** Federated credentials & configuration. */
-  @Prop() federated: FederatedConfig;
+  @Prop() federated: FederatedConfig = {};
   /** Passed from the Authenticator component in order to change Authentication state
    * e.g. SignIn -> 'Create Account' link -> SignUp
    */